+++ core/bin/dust_api.rs
@@ -19,13 +19,12 @@ use dust::{
     },
     databases::database::DatabaseRow,
     dataset,
-    http_utils::{error_response, APIError, APIResponse},
     project::{self},
     providers::provider::{provider, ProviderID},
     run,
     stores::postgres,
     stores::store,
-    utils,
+    utils::{self, error_response, APIError, APIResponse},
 };
 use hyper::http::StatusCode;
 use parking_lot::Mutex;
