+++ core/src/lib.rs
@@ -60,12 +60,13 @@ pub mod blocks {
     pub mod r#while;
 }
 
-pub mod sqlite_database;
-pub mod sqlite_workers;
+pub mod sqlite_workers {
+    pub mod sqlite_database;
+    pub mod sqlite_workers;
+}
 
 pub mod deno {
     pub mod script;
 }
 
 pub mod consts;
-pub mod http_utils;
